move methods into FutureExternalEvent
