+++ Revit.Async.Shared/Revit.Async.Shared.projitems
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Entities\DefaultResultHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\FutureExternalEvent.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensions\FutureExternalEventExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ExternalEventResultHandlerExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\SemaphoreSlimExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\TaskCompletionSourceExtensions.cs" />
